+++ RuntimeConsole/Other/CommandHandler/CommandAttribute.cs
@@ -9,18 +9,17 @@ namespace PotikotTools.Commands
     public class CommandAttribute : Attribute
     {
         public string Name;
+        public string Description;
         public bool IncludeTypeName;
 
-        public CommandAttribute(bool includeTypeName = true)
+        public CommandAttribute(string name, string description, bool includeTypeName = true)
         {
-            Name = string.Empty;
+            Name = name;
+            Description = description;
             IncludeTypeName = includeTypeName;
         }
 
-        public CommandAttribute(string name, bool includeTypeName = false)
-        {
-            Name = name;
-            IncludeTypeName = includeTypeName;
-        }
+        public CommandAttribute(string name, bool includeTypeName = true) : this(name, string.Empty, includeTypeName) { }
+        public CommandAttribute(bool includeTypeName = true) : this(string.Empty, string.Empty, includeTypeName) { }
     }
 }