+++ RuntimeConsole/Other/CommandHandler/CommandInfos/ActionCommandInfo.cs
@@ -7,18 +7,22 @@ namespace PotikotTools.Commands
         private Action _callback;
 
         public string Name { get; private set; }
-        public object Obj => null;
+        public string Description { get; private set; }
+        public object Context => null;
         public string HintText => Name;
 
         public Type[] ParameterTypes => null;
 
         public bool IsValid => _callback != null && !string.IsNullOrEmpty(Name);
 
-        public ActionCommandInfo(string name, Action callback)
+        public ActionCommandInfo(string name, string description, Action callback)
         {
             Name = name.Replace(' ', '_');
+            Description = description;
             _callback = callback;
         }
+
+        public ActionCommandInfo(string name, Action callback) : this(name, null, callback) { }
 
         public void Invoke(object[] parameters)
         {