+++ RuntimeConsole/Scripts/Editor/ConsolePreferencesSOEditor.cs
@@ -1,10 +1,10 @@
 ﻿using UnityEditor;
 using UnityEngine.UIElements;
 
-namespace PotikotTools.RuntimeConsole
+namespace PotikotTools.RuntimeConsole.Editor
 {
     [CustomEditor(typeof(ConsolePreferencesSO))]
-    public class ConsolePreferencesSOEditor : Editor
+    public class ConsolePreferencesSOEditor : UnityEditor.Editor
     {
         public override VisualElement CreateInspectorGUI() => new();
     }