+++ RuntimeConsole/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## [[1.0.2]](https://github.com/potikot/RuntimeConsole/tree/v1.0.2) - 
+
+- Added README Documentation
+- Optimized Command Searching
+- Added command descriptions
+- Added `help` command
+- Removed `enable` command
+- Removed `"Hide"` message on `ContextMenu.Hide()`
+
 ## [[1.0.1]](https://github.com/potikot/RuntimeConsole/tree/v1.0.1) - 2025-01-15
 
 - Added InitializeOnStart option